Pharmagenomic/requirements.txt
pytorch
streamlit
matplotlib
seaborn
scikit-learn